requirements.txt
# API
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1

# ML
scikit-learn==1.5.0
xgboost==2.0.3
pandas==2.2.2
numpy==1.26.4
imbalanced-learn==0.12.3

# MLflow
mlflow==2.13.0
psycopg2-binary==2.9.9

# DVC
dvc==3.51.2
dvc[gdrive]==3.51.2

# Monitoring
evidently==0.4.30

# Testing
pytest==8.2.0
pytest-cov==5.0.0
httpx==0.27.0

# Utils
python-dotenv==1.0.1
joblib==1.4.2
PyYAML==6.0.1
